# docs/openapi/driver.json
{
  "components": {
    "schemas": {
      "AdminDriverResponse": {
        "properties": {
          "accountStatus": {
            "type": "string"
          },
          "activeDeliveryId": {
            "type": "string"
          },
          "activeRouteId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentLat": {
            "type": "number"
          },
          "currentLng": {
            "type": "number"
          },
          "delivered": {
            "format": "int32",
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "failed": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "invitationExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "invitedByName": {
            "type": "string"
          },
          "lastInvitedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLocationAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "onlineStatus": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "registered": {
            "type": "boolean"
          },
          "suspendedReason": {
            "type": "string"
          },
          "totalDeliveries": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "InviteDriverRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "phone"
        ],
        "type": "object"
      },
...
      "StatsResponse": {
        "properties": {
          "cancelled": {
            "format": "int32",
            "type": "integer"
          },
          "delivered": {
            "format": "int32",
            "type": "integer"
          },
          "failed": {
            "format": "int32",
            "type": "integer"
          },
          "totalDeliveries": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "bearerFormat": "JWT",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "type": "http"
      },
      "Internal Secret": {
        "in": "header",
        "name": "X-Internal-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API documentation for the ASM Driver Service",
    "title": "Driver Service API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.1",
  "paths": {
    "/api/v1/admin/drivers": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminDriverResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List drivers",
        "tags": [
          "Admin Drivers"
        ]
      },
      "post": {
        "operationId": "invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDriverResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Invite driver — sends setup email, no password required",
        "tags": [
          "Admin Drivers"
        ]
      }
    },
...
    "/api/v1/driver/stats": {
      "get": {
        "description": "Returns the driver's own performance counters.",
        "operationId": "getStats",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get my delivery stats",
        "tags": [
          "Driver Operations"
        ]
      }
    },
...
  },
  "security": [
    {
      "Bearer Authentication": [],
      "Internal Secret": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://driver-service:8086"
    }
  ],
...
}